requirements.txt
django==4.0
djangorestframework==3.13.1
daphne==3.0.2
channels==3.0.4
channels-redis==3.3.1
django-cors-headers
aioredis
uvicorn
psycopg2
chanmqttproxy
unipath